Reconcile method

+++ server/src/main/java/com/soffid/iam/sync/service/ServerServiceImpl.java
@@ -2263,32 +2263,34 @@ public class ServerServiceImpl extends ServerServiceBase {
 		String t = tae.startVirtualSourceTransaction(true);
 		try {
 			getAccountService().updateAccount2(account);
-			List<RoleAccount> l = new LinkedList<RoleAccount>( grants );
-			List<RoleAccount> current = new LinkedList<RoleAccount>( getApplicationService().findRoleAccountByAccount(account.getId()) );
-			for (Iterator<RoleAccount> iterator = l.iterator(); iterator.hasNext();) {
-				RoleAccount ra = iterator.next();
-				boolean found = false;
-				for (Iterator<RoleAccount> iterator2 = current.iterator(); iterator2.hasNext();) {
-					RoleAccount ra2 = iterator2.next();
-					if (ra2.getRoleName().equals(ra.getRoleName())) {
-						if (ra2.getDomainValue() == null ||
-								ra2.getDomainValue().getValue() == null ||
-								ra.getDomainValue() != null && ra2.getDomainValue().getValue().equals(ra.getDomainValue().getValue())) {
-							found = true;
-							iterator2.remove();
-							iterator.remove();
-							break;
+			if (grants != null) {
+				List<RoleAccount> l = new LinkedList<RoleAccount>( grants );
+				List<RoleAccount> current = new LinkedList<RoleAccount>( getApplicationService().findRoleAccountByAccount(account.getId()) );
+				for (Iterator<RoleAccount> iterator = l.iterator(); iterator.hasNext();) {
+					RoleAccount ra = iterator.next();
+					boolean found = false;
+					for (Iterator<RoleAccount> iterator2 = current.iterator(); iterator2.hasNext();) {
+						RoleAccount ra2 = iterator2.next();
+						if (ra2.getRoleName().equals(ra.getRoleName())) {
+							if (ra2.getDomainValue() == null ||
+									ra2.getDomainValue().getValue() == null ||
+									ra.getDomainValue() != null && ra2.getDomainValue().getValue().equals(ra.getDomainValue().getValue())) {
+								found = true;
+								iterator2.remove();
+								iterator.remove();
+								break;
+							}
 						}
 					}
+					if (!found) {
+						getApplicationService().create(ra);
+					}
 				}
-				if (!found) {
-					getApplicationService().create(ra);
+				
+				for (Iterator<RoleAccount> iterator2 = current.iterator(); iterator2.hasNext();) {
+					RoleAccount ra2 = iterator2.next();
+					getApplicationService().delete(ra2);
 				}
-			}
-			
-			for (Iterator<RoleAccount> iterator2 = current.iterator(); iterator2.hasNext();) {
-				RoleAccount ra2 = iterator2.next();
-				getApplicationService().delete(ra2);
 			}
 		} finally {
 			tae.finishVirtualSourceTransaction(t);
